+++ realtor/realtor/urls.py
@@ -20,6 +20,5 @@ urlpatterns = [
     path("admin/", admin.site.urls),
     path("properties/", include("properties.urls")),
     path("users/", include("users.urls")),
-    path("agents/", include("transactions.urls")),
     path("authentication/", include("authentication.urls")),
 ]
