+++ realtor/users/admin.py
@@ -1,7 +1,7 @@
 from django.contrib import admin
 from rangefilter.filter import DateRangeFilter
 
-from .models import Application, PreferredProperty, User
+from .models import Agent, Application, Contract, PreferredProperty, User
 
 # Register your models here.
 
@@ -40,3 +40,30 @@ class PreferredPropertyAdmin(admin.ModelAdmin):
     @admin.display(description="Address")
     def full_address(self, obj):
         return obj.city + " " + obj.district + " " + obj.zone
+
+
+@admin.register(Agent)
+class AgentAdmin(admin.ModelAdmin):
+
+    list_display = ["id", "user", "rating"]
+    list_per_page = 5
+    list_filter = ["rating"]
+
+
+@admin.register(Contract)
+class ContractAdmin(admin.ModelAdmin):
+
+    list_display = [
+        "id",
+        "mortgage_ratio",
+        "short_agreements",
+        "start_date",
+        "end_date",
+    ]
+    list_per_page = 5
+    list_filter = (("start_date", DateRangeFilter), ("end_date", DateRangeFilter), "mortgage_ratio")
+    search_fields = ["mortgage_ratio"]
+
+    @admin.display(description="brief agreements")
+    def short_agreements(self, obj):
+        return obj.agreements[:20]
